fix: Addressing review comments and code refactoring
