+++ hedera-mirror-web3/src/main/java/com/hedera/mirror/web3/evm/contracts/execution/MirrorEvmMessageCallProcessor.java
@@ -18,18 +18,12 @@ package com.hedera.mirror.web3.evm.contracts.execution;
 
 import static com.hedera.node.app.service.evm.contracts.operations.HederaExceptionalHaltReason.FAILURE_DURING_LAZY_ACCOUNT_CREATE;
 import static com.hedera.services.store.contracts.precompile.codec.DecodingFacade.wrapUnsafely;
-import static org.apache.tuweni.bytes.Bytes.EMPTY;
 import static org.hyperledger.besu.evm.frame.ExceptionalHaltReason.INSUFFICIENT_GAS;
-import static org.hyperledger.besu.evm.frame.MessageFrame.State.COMPLETED_SUCCESS;
 import static org.hyperledger.besu.evm.frame.MessageFrame.State.EXCEPTIONAL_HALT;
-import static org.hyperledger.besu.evm.frame.MessageFrame.State.REVERT;
 
 import com.hedera.mirror.web3.evm.config.PrecompiledContractProvider;
 import com.hedera.mirror.web3.evm.store.contract.EntityAddressSequencer;
 import com.hedera.mirror.web3.evm.store.contract.HederaEvmStackedWorldStateUpdater;
-import com.hedera.node.app.service.evm.contracts.execution.HederaEvmMessageCallProcessor;
-import com.hedera.node.app.service.evm.store.contracts.AbstractLedgerEvmWorldUpdater;
-import com.hedera.node.app.service.evm.store.contracts.precompile.EvmHTSPrecompiledContract;
 import com.hedera.services.contracts.gascalculator.GasCalculatorHederaV22;
 import com.hedera.services.ledger.BalanceChange;
 import com.hedera.services.txns.crypto.AbstractAutoCreationLogic;
@@ -38,6 +32,7 @@ import com.hederahashgraph.api.proto.java.AccountAmount;
 import com.hederahashgraph.api.proto.java.AccountID;
 import com.hederahashgraph.api.proto.java.ResponseCodeEnum;
 import com.hederahashgraph.api.proto.java.Timestamp;
+import com.swirlds.base.utility.Pair;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Optional;
 import org.apache.tuweni.bytes.Bytes;
@@ -49,7 +44,7 @@ import org.hyperledger.besu.evm.precompile.PrecompileContractRegistry;
 import org.hyperledger.besu.evm.precompile.PrecompiledContract;
 import org.hyperledger.besu.evm.tracing.OperationTracer;
 
-public class MirrorEvmMessageCallProcessor extends HederaEvmMessageCallProcessor {
+public class MirrorEvmMessageCallProcessor extends AbstractEvmMessageCallProcessor {
     private final AbstractAutoCreationLogic autoCreationLogic;
     private final EntityAddressSequencer entityAddressSequencer;
 
@@ -70,38 +65,12 @@ public class MirrorEvmMessageCallProcessor extends HederaEvmMessageCallProcessor
     @Override
     protected void executeHederaPrecompile(
             PrecompiledContract contract, MessageFrame frame, OperationTracer operationTracer) {
-        Bytes output = EMPTY;
-        Long gasRequirement = 0L;
-        if (contract instanceof EvmHTSPrecompiledContract htsPrecompile) {
-            var updater = (AbstractLedgerEvmWorldUpdater) frame.getWorldUpdater();
-            final var costedResult = htsPrecompile.computeCosted(
-                    frame.getInputData(),
-                    frame,
-                    (now, minimumTinybarCost) -> minimumTinybarCost,
-                    updater.tokenAccessor());
-            output = costedResult.getValue();
-            gasRequirement = costedResult.getKey();
-        }
-        if (!"HTS".equals(contract.getName()) && !"EvmHTS".equals(contract.getName())) {
-            output = contract.computePrecompile(frame.getInputData(), frame).getOutput();
-            gasRequirement = contract.gasRequirement(frame.getInputData());
-        }
+        Pair<Long, Bytes> costAndResult = calculatePrecompileGasAndOutput(contract, frame);
 
-        operationTracer.tracePrecompileCall(frame, gasRequirement, output);
-        if (frame.getState() == REVERT) {
-            return;
-        }
-        if (frame.getRemainingGas() < gasRequirement) {
-            frame.decrementRemainingGas(frame.getRemainingGas());
-            frame.setExceptionalHaltReason(Optional.of(INSUFFICIENT_GAS));
-            frame.setState(EXCEPTIONAL_HALT);
-        } else if (output != null) {
-            frame.decrementRemainingGas(gasRequirement);
-            frame.setOutputData(output);
-            frame.setState(COMPLETED_SUCCESS);
-        } else {
-            frame.setState(EXCEPTIONAL_HALT);
-        }
+        Long gasRequirement = costAndResult.left();
+        Bytes output = costAndResult.right();
+
+        traceAndHandleExecutionResult(frame, operationTracer, gasRequirement, output);
     }
 
     /**
